rework Registration send activation email add several migration to DB add profile page

+++ src/main/java/com/example/sweater/Model/Message.java
@@ -6,7 +6,7 @@ import javax.persistence.*;
 public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer Id;
+    private Long Id;
 
     private String text;
     private String tag;
@@ -60,7 +60,7 @@ public class Message {
         this.tag = tag;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return Id;
     }
